requirements.txt
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
requests==2.28.1
jsonpath==0.82
pytest==7.1.2